requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0
playwright>=1.40.0
urllib3>=2.0.0
PyGithub>=2.0.0
